replace muls with shifts
but the muls exist in the Concat3 version but are much less expensive?
I think it's because it is able to fold the muls across the different calls? it seems to only do one mul per memcpy
or maybe it's (additionally) because the start of the strings are always 0?
oh, no it does two imuls per copy?
oh, no. it seems to be saving the multiplication result.